Implement ETAG caching and checking.
Caching server now checks orign server responses for an `ETAG` header, storing the result into a `.etag` file next to the actual cached content.  When forwarding users to this cached copy, the caching server now pings for the `ETAG` via a `HEAD` request to the origin server before forwarding.  This ensures that the file has not changed, and doesn't seem to slow it down much (~160ms measured for `codeload.github.com`).  If it can't get the current `ETAG`, it goes ahead and serves its cached file.  If the ETAG doesn't match, it deletes the file and downloads the new one, making a note of it in the log file.